+++ src/main/java/svenhjol/charm/module/portable_crafting/network/ClientSendOpenCrafting.java
@@ -7,4 +7,9 @@ import svenhjol.charm.network.Id;
  * Client sends empty message to ask server to open the crafting container.
  */
 @Id("charm:open_crafting")
-public class ClientSendOpenCrafting extends ClientSender { }
+public class ClientSendOpenCrafting extends ClientSender {
+    @Override
+    protected boolean showDebugMessages() {
+        return false;
+    }
+}