+++ src/main/java/svenhjol/charm/module/portable_crafting/network/ServerReceiveOpenCrafting.java
@@ -19,4 +19,9 @@ public class ServerReceiveOpenCrafting extends ServerReceiver {
             PortableCrafting.openContainer(player);
         });
     }
+
+    @Override
+    protected boolean showDebugMessages() {
+        return false;
+    }
 }